Switches to basing site off of html5 boilerplate.
Everything is a mess right now. There are still older css files around ( but unused ),
most of the index page is borked and a variety of other problems but there is some initial
progress.

This is an intermediate commit and no one should ever pull this out and look at it, least their eyes
burn.