Changes logo/banner style.
Switches to just have a logo/banner for redline that runs across the page at the top horizontally.
Previous version wrapped around left side and was a problem with fluid layouts.
Will possibly revisit later. Moving ahead with this simple design for now.

Site nav needs to be adjust as it doesn't look good right now.